+++ pkg/providers/property/data_akamai_property_rules_builder_test.go
@@ -61,18 +61,28 @@ func TestDataPropertyRulesBuilder(t *testing.T) {
 			})
 		})
 	})
-}
-
-func testCheckResourceAttrJSONFromFixture(name, key, fixturePath string) func(s *terraform.State) error {
-	value := loadFixtureString(fixturePath)
-	checkFn := testCheckResourceAttrJSON(name, key, value)
-
-	return func(s *terraform.State) error {
-		if err := checkFn(s); err != nil {
-			return fmt.Errorf("expected from: %s, %w", fixturePath, err)
-		}
-		return nil
-	}
+	t.Run("fails on rule with is_secure outside default rule", func(t *testing.T) {
+		useClient(nil, nil, func() {
+			resource.UnitTest(t, resource.TestCase{
+				ProviderFactories: testAccProviders,
+				Steps: []resource.TestStep{{
+					Config:      loadFixtureString("testdata/TestDSPropertyRulesBuilder/rules_with_is_secure_outside_default.tf"),
+					ExpectError: regexp.MustCompile(`cannot be used outside 'default' rule: is_secure`),
+				}},
+			})
+		})
+	})
+	t.Run("fails on rule with variable outside default rule", func(t *testing.T) {
+		useClient(nil, nil, func() {
+			resource.UnitTest(t, resource.TestCase{
+				ProviderFactories: testAccProviders,
+				Steps: []resource.TestStep{{
+					Config:      loadFixtureString("testdata/TestDSPropertyRulesBuilder/rules_with_variable_outside_default.tf"),
+					ExpectError: regexp.MustCompile(`cannot be used outside 'default' rule: variable`),
+				}},
+			})
+		})
+	})
 }
 
 func testCheckResourceAttrJSON(name, key, value string) func(s *terraform.State) error {